pyproject.toml
[tool.poetry]
name = "somenergia-modbus-reader"
version = "0.1.0"
description = "A lightweight modbus reader for somenergia's plants"
authors = ["Equip de Dades de SomEnergia <equip.dades@somenergia.coop>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
typer = {extras = ["all"], version = "^0.6.1"}
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4.42"
psycopg2 = "^2.9.4"
pymodbus = "^3.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
mkdocs-material = "^8.5.6"
ipdb = "^0.13.9"
pymdown-extensions = "^9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
